complete limited access

+++ LuxeAssist/request/views.py
@@ -9,7 +9,10 @@ from main.models import Payment
 
 
 def add_Request_view(request:HttpRequest, service_id):
-   
+
+    if not request.user.is_authenticated:
+        return render(request, "main/user_not_found.html", status=401)
+
     service = Service.objects.get(id = service_id)  
     if request.method == "POST":
         newRequest = Request(user = request.user, service = service, note = request.POST["note"], date = request.POST["date"], request_price = request.POST["request_price"])
@@ -18,8 +21,10 @@ def add_Request_view(request:HttpRequest, service_id):
         
 
 def user_requests_view(request: HttpRequest):
+
+    if not request.user.is_authenticated :
+       return render(request,"main/user_not_found.html", status=401)
     try:
-        
         requests = Request.objects.filter(user= request.user)
 
         for index, r in enumerate(requests):
@@ -33,8 +38,7 @@ def user_requests_view(request: HttpRequest):
     
 
 def concierge_requests_view(request: HttpRequest): 
-
-   
+   # try:
     services=Service.objects.filter(user=request.user)
     requests = Request.objects.filter(service__in= services)
 
@@ -94,13 +98,15 @@ def chekeUser_unPayment_view(request: HttpRequest):
     
 
 def admin_requests_view(request: HttpRequest):
+
+    if not request.user.is_superuser :
+       return render(request,"main/user_not_found.html", status=401)
    try: 
       requests = Request.objects.order_by('-date')
      
      
       return render(request ,"request/admin_requests_view.html" , {"requests" : requests})
    except Exception as e:
-
       return render(request, "main/user_not_found.html")
    
 
@@ -122,6 +128,9 @@ def request_details_view(request: HttpRequest,requsets_id):
      
 
 def cancel_request_view(request: HttpRequest, requset_id):
+    if request.user.is_superuser and request.user.is_staff :
+       return render(request,"main/user_not_found.html", status=401)
+    
     try:
         requests = Request.objects.get(id = requset_id)
         requests.status = "Cancle"
@@ -134,7 +143,11 @@ def cancel_request_view(request: HttpRequest, requset_id):
 
 def update_price_view(request:HttpRequest, requests_id):
 
-    requests = Request.objects.get(id = requests_id)  
+    requests = Request.objects.get(id = requests_id) 
+
+    if request.user.is_superuser and request.user.is_staff :
+        return render(request,"main/user_not_found.html", status=401)
+     
     if request.method == "POST":
         requests.request_price = request.POST["request_price"]
         requests.save()
@@ -142,7 +155,11 @@ def update_price_view(request:HttpRequest, requests_id):
 
 
 def add_status_view(request: HttpRequest, requests_id):
-    requests = Request.objects.get(id = requests_id)  
+    if not request.user.is_authenticated and not request.user.groups.exists :
+       return render(request,"main/user_not_found.html", status=401)
+
+    requests = Request.objects.get(id = requests_id)
+
     if request.method == "POST":
         requests.status = request.POST["status"]
         requests.save()
@@ -158,14 +175,16 @@ def new_requestConcierge_view(request: HttpRequest , requset_id):
         return render(request, "main/user_not_found.html")
 
 def delete_request_admin_view(request: HttpRequest, requset_id):
+
+    if not request.user.is_superuser :
+        return render(request,"main/user_not_found.html", status=401)
     try:
         requsets=Request.objects.get(id=requset_id)
         requsets.delete()
         return redirect("request:admin_requests_view")
     except  Exception as e:
 
-        return render(request, "main/user_not_found.html")
-    
+        return render(request, "main/user_not_found.html") 
 
 
 def requests_status_viwe(request:HttpRequest,item):
